Bugfixes, API updates, more tests
- The job runner script now requires the service version supplied in the
RunJobParameters to be a git commit. Since that's what the NJSW
supplies, that's OK.

- Fixed a bug where RPC errors with only a name field would cause an
"unknown error" in the logs.

- Updated the get_provenance call in the callback server for thread
safety.

- Added a synchronized block in the callback server run method to ensure
the integrity of the cached module versions.

- The NJSW now checks that the ModuleVersionInformation is available
from the catalog in all cases and so fails early instead of starting a
job that is destined to fail.

- Improved error handling for module lookups in the
SubsequentCallRunner.

- Added a number of tests for these changes and a few tests for the
previous changes.